requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-cloud-firestore==2.16.0
google-auth==2.29.0
pydantic==1.10.15
python-dotenv==1.0.1
